requirements.txt
langchain
langchain-openai
langchain-community
langchainhub
huggingface_hub
openai
python-dotenv
unstructured
sentence-transformers
transformers
tokenizers
opensearch-py
python-dotenv